[REF] refactor test by using odoo-test-help lib
In order to avoid duplicated code between projet
The odoo-test-lib include an helper to load fake models
use it
