+++ .isort.cfg
@@ -9,4 +9,4 @@ line_length=88
 known_odoo=odoo
 known_odoo_addons=odoo.addons
 sections=FUTURE,STDLIB,THIRDPARTY,ODOO,ODOO_ADDONS,FIRSTPARTY,LOCALFOLDER
-known_third_party =mock,setuptools,urlparse,vcr_unittest
+known_third_party =mock,odoo_test_helper,setuptools,urlparse,vcr_unittest
