+++ .travis.yml
@@ -36,7 +36,7 @@ jobs:
 
 env:
   global:
-  - VERSION="10.0" TESTS="0" LINT_CHECK="0" MAKEPOT="0"
+  - VERSION="10.0" TESTS="0" LINT_CHECK="0" MAKEPOT="0" MQT_DEP="PIP"
 
 install:
   - git clone --depth=1 https://github.com/OCA/maintainer-quality-tools.git ${HOME}/maintainer-quality-tools
